editor/libeditor/tests/test_bug966552.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=966552
-->
<head>
  <title>Test for Bug 966552</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=966552">Mozilla Bug 966552</a>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
<script class="testbody" type="text/javascript">
SimpleTest.waitForExplicitFinish();

var testWindow = window.open("file_bug966552.html", "", "test-966552");
testWindow.addEventListener("load", function() {
  runTest(testWindow);
}, {once: true});

function runTest(win) {
  SimpleTest.waitForFocus(function() {
    var doc = win.document;
    var sel = win.getSelection();
    doc.body.focus();
    sel.collapse(doc.body.firstChild, 2);
    synthesizeKey("VK_BACK_SPACE", {ctrlKey: true}, win);
    is(doc.body.textContent.trim(), "st");
    win.close();
    SimpleTest.finish();
  }, win);
}

</script>
</pre>
</body>
</html>
